requirements.txt
# 스키마 추출 (extract-schema.py)
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# 스키마 임베딩 (schema_index.py) - TEI API 사용
numpy>=1.24.0

# 프롬프트 템플릿
jinja2>=3.1.0

# HTTP LLM 호출
requests>=2.31.0
